put iterators back into a separate files
