+++ src/main.js
@@ -65,7 +65,11 @@ require("./generator.js");
 require("./file.js");
 require("./ffi.js");
 require("./range.js");
-require("./iteratorobjects.js");
+require("./enumerate.js");
+require("./filter.js");
+require("./map.js");
+require("./reversed.js");
+require("./zip.js");
 require("./token.js");
 require("./tokenize.js");
 require("../gen/parse_tables.js");
